Refactor: Remove deprecated insertContent and insertLineByLine utilities; introduce saveFiles utility
