+++ src/utils/index.ts
@@ -34,8 +34,6 @@ export * from './getUserInput';
 export * from './getVariables';
 export * from './getWebviewUrl';
 export * from './htmlDecode';
-export * from './insertContent';
-export * from './insertLineByLine';
 export * from './insertVariables';
 export * from './isProjectInitialized';
 export * from './lowercaseFirstLetter';
@@ -43,8 +41,8 @@ export * from './parseWinPath';
 export * from './readFile';
 export * from './removeDemoDuplicates';
 export * from './removeDemosForCurrentPosition';
-export * from './replaceContent';
 export * from './sanitizeFileName';
+export * from './saveFiles';
 export * from './setContext';
 export * from './sleep';
 export * from './sortFiles';
