+++ src/utils/saveFiles.ts
@@ -1,0 +1,5 @@
+import { commands } from 'vscode';
+
+export const saveFiles = async (): Promise<void> => {
+  await commands.executeCommand('workbench.action.files.save');
+};
